refactor(cmd): split endpoint.go into concern-focused files
Move cohesive subsystems out of the 1500-line endpoint.go into their own files
within package cmd, with no behavior change:

- endpoint_hooks.go: hook command vars and install/uninstall/status helpers.
- endpoint_install.go: install/status/discover/uninstall/repair/dashboard
  command vars and run helpers.
- endpoint_elastic.go: the local Elasticsearch dev-stack up/down helpers.

endpoint.go shrinks from ~1500 to ~700 lines and now holds the root command,
shared options, init() wiring, and shared helpers. This is a pure code move:
the set of top-level declarations is unchanged and generated command docs are
byte-identical to baseline.
